Reduce memory usage to fix broken builds (#8)
Recent indexing pipeline workflows were broken because of the
out-of-disk-space error. This was caused by Debug info taking too much
space and the fact that some generated files depend on actual libraries
and binaries which blows up the build size. Eventually we would like to
fix these generated files dependencies as they are unlikely to actually
need these libraries. However, we have no control over other projects'
dependency graphs and they can introduce some regresseions by depending
on some of LLVM/Clang libraries, too.

Hence, the solution is to

1. Build in Release mode: that reduces build size on disk and also makes
the build step faster. We only care about generated files here and they
are unlikely to differ between various build types.
2. Generate compile_commands.json using the Debug build configuration
and feed them to clangd-indexer. All the sources are generated in the
previous step and we are not running any build commands after CMake
regeneration. Having it in the same directory ensures the indexer can
find the needed files.